refactor: client, remote type 구분 및 재정의
