+++ frontend/src/types/client/user.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: number;
+  email: string;
+  nickname: string;
+}
