--- frontend/src/utils/constants/kkogkkog.ts
@@ -1,4 +1,0 @@
-export const THUMBNAIL = {
-  COFFEE: '/assets/images/coffee.png',
-  DRINK: '/assets/images/beer.png',
-} as const;
